fix: optimize sample app code
